server_archivos/rest/v1/repositorio/routers.py
from rest_framework.routers import DefaultRouter

from . import views

router = DefaultRouter()

router.register(
    prefix="software",
    viewset=views.SoftwareViewSet,
    basename="software",
)

router.register(
    prefix="carpeta",
    viewset=views.CarpetaViewSet,
    basename="carpeta",
)
