requirements/base.txt
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
python-decouple==3.8
django-debug-toolbar==4.2
psycopg2-binary==2.9.7
gunicorn==20.1.0